SMS-3864: Regenerate all apps with latest tools versions.

+++ apps/dual_core_demo/firmware/src/config/DualCoreDemo_Core1/peripheral/clk/plib_clk.c
@@ -21,7 +21,6 @@
 * THAT YOU HAVE PAID DIRECTLY TO MICROCHIP FOR THIS SOFTWARE.
 *******************************************************************************/
 #include <stdbool.h>
-#include "device.h"
 #include "plib_clk.h"
 /*********************************************************************************
                         Check Peripheral clock status
@@ -40,12 +39,10 @@ static bool PeripheralClockStatus(uint32_t periph_id)
     {
         status = (*(volatile uint32_t* const)((PMC_BASE_ADDRESS +
                                                         csr_offset[index])));
-        retval = ((status & (1 << (periph_id % 32U))) != 0U);
+        retval = ((status & ((uint32_t)1U << (periph_id % 32U))) != 0U);
     }
     return retval;
 }
-
-
 /*********************************************************************************
                         Initialize Peripheral clocks
 *********************************************************************************/
